Synchronize database migrations with production schema
- Created missing migrations for 5 critical tables:
  • evaluations: Core business table for invoices/evaluations
  • attributables: Polymorphic attributes system
  • purchaser_special_attributes: Purchaser configurations
  • user_dates: User date tracking
  • attributes: Legacy attributes data

- Fixed foreign key type inconsistencies (unsignedInteger → unsignedBigInteger)
- Removed duplicate migrations that were causing conflicts
- Consolidated migrations from old_migrations directories
- All 98 migrations now run successfully on clean database

+++ database/migrations/2024_03_23_161628_add_user_id_field_to_locations_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('locations', function (Blueprint $table) {
-            $table->unsignedInteger("user_id")->nullable();
+            $table->unsignedBigInteger("user_id")->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
